+++ tests/error/derives_for_enums_with_source.rs
@@ -12,6 +12,10 @@ enum TestErr {
         source: SimpleErr,
         field: i32,
     },
+    NamedImplicitBoxedSource {
+        source: Box<dyn Error + Send + 'static>,
+        field: i32,
+    },
     NamedExplicitNoSource {
         #[error(not(source))]
         source: SimpleErr,
@@ -87,6 +91,17 @@ fn named_implicit_no_source() {
 fn named_implicit_source() {
     let err = TestErr::NamedImplicitSource {
         source: SimpleErr,
+        field: 0,
+    };
+
+    assert!(err.source().is_some());
+    assert!(err.source().unwrap().is::<SimpleErr>());
+}
+
+#[test]
+fn named_implicit_boxed_source() {
+    let err = TestErr::NamedImplicitBoxedSource {
+        source: Box::new(SimpleErr),
         field: 0,
     };